+++ CHANGELOG.md
@@ -15,6 +15,7 @@
     + `NSDateFormatter (ESAdditions)` category
     + `-[NSUserDefaults setObject:forKeyPath:]`
     + `ESNetworkHelper` class
+    + `-[UIView snapshotViewAfterScreenUpdates:]`
 - Renamed:
     + `ES_STOPWATCH_BEGIN` `ES_STOPWATCH_END` => `ESBenchmark()`
     + `ESDispatchOnMainThreadAsynchrony` => `es_dispatch_async_main`